# tauri/packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Springtale management API",
    "description": "Local-first bot automation. Binds loopback; bearer-authenticated.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/drift/recipe/{id}": {
      "get": {
        "tags": [
          "drift"
        ],
        "summary": "GET /drift/recipe/{id} — drift window for one recipe.",
        "operationId": "drift_recipe",
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "description": "Filter to one bot's runs. Both this and `formation_id`\ncan be set — they intersect.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "formation_id",
            "in": "query",
            "description": "Filter to one formation's runs.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "rule_id",
            "in": "query",
            "description": "Filter to one rule's runs. Required for `rule_drift`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "recent_n",
            "in": "query",
            "description": "Most-recent run count. Default 10.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "baseline_n",
            "in": "query",
            "description": "Older runs preceding `recent_n`. Default 30.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe drift report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "GET /recipes — list with optional filter via query params.",
        "operationId": "recipes_list",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/RecipeCategory"
                }
              ]
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "favorites_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/RecipeSort"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching recipes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes/import": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/import — import a recipe from raw TOML.",
        "operationId": "recipes_import_toml",
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Imported recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/user": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/user — save to the user library. Runs Clear Check.",
        "operationId": "recipes_save_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recipe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved user recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/user/{id}": {
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "DELETE /recipes/user/{id} — remove a user recipe.",
        "operationId": "recipes_delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "GET /recipes/{id} — fetch one. 404 if not found.",
        "operationId": "recipes_get_one",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/apply": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/apply — W1.C deploy with user inputs.",
        "operationId": "recipes_apply",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyReport"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/export": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "GET /recipes/{id}/export — TOML for sharing.",
        "operationId": "recipes_export_toml",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe TOML",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/favorite": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/favorite — toggle a recipe in/out of favorites.",
        "operationId": "recipes_toggle_favorite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New favorite state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/fork": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/fork — duplicate into the user library.",
        "operationId": "recipes_fork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Forked recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/pieces": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "GET /recipes/{id}/pieces — W2.D borrowable slots.",
        "operationId": "recipes_list_pieces",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe pieces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipePieceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/preflight": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/preflight — W1.D Deploy-readiness checklist.",
        "operationId": "recipes_preflight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Preflight report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReport"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/preview": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/preview — W2.C dry-run with comic-strip narrative.",
        "operationId": "recipes_preview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Preview report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewReport"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/recent": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/recent — push onto the recently-used list.",
        "operationId": "recipes_record_recent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recorded"
          }
        }
      }
    },
    "/recipes/{id}/render": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/render — show-as-code disclosure renders the\nassembled TOML for the user's current inputs. Read-only; takes\nthe same payload as `/apply` but doesn't materialise.",
        "operationId": "recipes_render",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rendered rule TOML",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/test-step": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "POST /recipes/{id}/test-step — Phase C \"Test This Step\" dry run\nthrough `step_index` of `rule_index`.",
        "operationId": "recipes_test_step",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Recipe id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestStepBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Step test report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestStepReport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriftReport": {
        "type": "object",
        "description": "Aggregate drift signal for a recipe / rule.",
        "required": [
          "recent_runs",
          "baseline_runs",
          "latency",
          "success_rate",
          "refusal_rate",
          "overall"
        ],
        "properties": {
          "baseline_runs": {
            "type": "integer",
            "format": "int32",
            "description": "How many baseline runs informed the analysis.",
            "minimum": 0
          },
          "latency": {
            "$ref": "#/components/schemas/LatencyDrift",
            "description": "Latency drift — median + p95 duration_ms delta."
          },
          "overall": {
            "$ref": "#/components/schemas/DriftClass",
            "description": "Overall verdict — the worst-of the three signals."
          },
          "recent_runs": {
            "type": "integer",
            "format": "int32",
            "description": "How many recent runs informed the analysis.",
            "minimum": 0
          },
          "refusal_rate": {
            "$ref": "#/components/schemas/RateDrift",
            "description": "Refusal-rate drift (fraction of `extract` steps with\n`schema_invalid` / `refused` errors)."
          },
          "success_rate": {
            "$ref": "#/components/schemas/RateDrift",
            "description": "Success-rate drift (fraction of `succeeded` rows)."
          }
        }
      },
...
      "FavoriteResponse": {
        "type": "object",
        "required": [
          "recipe_id",
          "now_favorite"
        ],
        "properties": {
          "now_favorite": {
            "type": "boolean"
          },
          "recipe_id": {
            "type": "string"
          }
        }
      },
...
      "RecipeCategory": {
        "type": "string",
        "enum": [
          "messaging",
          "coding",
          "web",
          "ai_assistant",
          "daily",
          "safety_privacy",
          "custom"
        ]
      },
...
      "RecipeInputs": {
        "type": "object",
        "description": "What the user submitted when picking a recipe.",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "Map from `InputField.id` → user-supplied value.",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "RecipeSort": {
        "type": "string",
        "enum": [
          "recommended",
          "name",
          "recent"
        ]
      },
...
    }
  },
...
}